Create first SettingsView and Initial Streaming Service to Preview Music
For the current moment the Preview only works by using Apple Music.
In an additional step, there will be a Repository which takes up the job
of using AppleMusicService and SpotifyService. If AppleMusicService is
not available or not able to fetch/play songs, then it will take
Spotify. This setup still needs to be implemented. Also the GameEngine
is not yet there.